fix: share TOML-safe catalog transport gate
