+++ src/config-manager.mjs
@@ -43,6 +43,7 @@ import {
   SIGNED_PROVIDER_MODE_PATH,
   loopback,
 } from "./paths.mjs";
+import { scanTomlDocument } from "./toml-structure.mjs";
 
 const legacyRouterBaseUrl = loopback(PORTS.router, "/v1");
 const startMarker = "# BEGIN codex-router-managed";
@@ -438,239 +439,8 @@ function replaceRootValue(contents, key, value) {
     .trimEnd();
 }
 
-function tomlBasicKey(raw) {
-  let decoded = "";
-  for (let index = 0; index < raw.length; index += 1) {
-    const character = raw[index];
-    if (character !== "\\") {
-      decoded += character;
-      continue;
-    }
-    const escape = raw[++index];
-    const simple = {
-      b: "\b",
-      t: "\t",
-      n: "\n",
-      f: "\f",
-      r: "\r",
-      '"': '"',
-      "\\": "\\",
-    }[escape];
-    if (simple !== undefined) {
-      decoded += simple;
-      continue;
-    }
-    const digits = escape === "u" ? 4 : escape === "U" ? 8 : 0;
-    const code = digits ? raw.slice(index + 1, index + 1 + digits) : "";
-    if (!digits || !new RegExp(`^[0-9a-fA-F]{${digits}}$`).test(code)) {
-      return undefined;
-    }
-    const value = Number.parseInt(code, 16);
-    if (value > 0x10ffff || (value >= 0xd800 && value <= 0xdfff)) return undefined;
-    decoded += String.fromCodePoint(value);
-    index += digits;
-  }
-  return decoded;
-}
-
-function tomlDottedKey(raw) {
-  const parts = [];
-  let index = 0;
-  while (index < raw.length) {
-    while (/\s/.test(raw[index] || "")) index += 1;
-    if (index >= raw.length) return undefined;
-    const quote = raw[index] === '"' || raw[index] === "'" ? raw[index++] : undefined;
-    let value = "";
-    if (quote) {
-      let closed = false;
-      while (index < raw.length) {
-        if (raw[index] === quote) {
-          index += 1;
-          closed = true;
-          break;
-        }
-        if (quote === '"' && raw[index] === "\\") {
-          const escapeStart = index;
-          index += 2;
-          if (raw[escapeStart + 1] === "u") index += 4;
-          else if (raw[escapeStart + 1] === "U") index += 8;
-          value += raw.slice(escapeStart, index);
-        } else {
-          value += raw[index++];
-        }
-      }
-      if (!closed) return undefined;
-      if (quote === '"') value = tomlBasicKey(value);
-      if (value === undefined) return undefined;
-    } else {
-      const start = index;
-      while (index < raw.length && raw[index] !== "." && !/\s/.test(raw[index])) {
-        index += 1;
-      }
-      value = raw.slice(start, index);
-      if (!value || !/^[A-Za-z0-9_-]+$/.test(value)) return undefined;
-    }
-    parts.push(value);
-    while (/\s/.test(raw[index] || "")) index += 1;
-    if (index === raw.length) return parts;
-    if (raw[index] !== ".") return undefined;
-    index += 1;
-  }
-  return undefined;
-}
-
-function ambiguousTomlBoundary(line, detail) {
-  throw new Error(
-    `Refusing ambiguous TOML table boundaries at line ${line}: ${detail}.`,
-  );
-}
-
-function tableHeaderAtLine(line, lineNumber) {
-  let index = 0;
-  while (/\s/.test(line[index] || "")) index += 1;
-  if (line[index] !== "[") return undefined;
-  const array = line[index + 1] === "[";
-  index += array ? 2 : 1;
-  const start = index;
-  let quote;
-  let escaped = false;
-  let end = -1;
-  while (index < line.length) {
-    const character = line[index];
-    if (quote === '"') {
-      if (escaped) escaped = false;
-      else if (character === "\\") escaped = true;
-      else if (character === quote) quote = undefined;
-      index += 1;
-      continue;
-    }
-    if (quote === "'") {
-      if (character === quote) quote = undefined;
-      index += 1;
-      continue;
-    }
-    if (character === '"' || character === "'") {
-      quote = character;
-      index += 1;
-      continue;
-    }
-    if (character === "[") {
-      ambiguousTomlBoundary(lineNumber, "an unquoted opening bracket appears in a table header");
-    }
-    if (character === "]") {
-      if (array && line[index + 1] !== "]") {
-        ambiguousTomlBoundary(lineNumber, "an array-table header has only one closing bracket");
-      }
-      end = index;
-      index += array ? 2 : 1;
-      break;
-    }
-    index += 1;
-  }
-  if (quote) ambiguousTomlBoundary(lineNumber, "a quoted table key is unterminated");
-  if (end === -1) ambiguousTomlBoundary(lineNumber, "a table header is unterminated");
-  const remainder = line.slice(index).trimStart();
-  if (remainder && !remainder.startsWith("#")) {
-    ambiguousTomlBoundary(lineNumber, "unexpected text follows a table header");
-  }
-  const path = tomlDottedKey(line.slice(start, end).trim());
-  if (!path) ambiguousTomlBoundary(lineNumber, "the table key cannot be decoded safely");
-  return path;
-}
-
-// This is deliberately a boundary lexer rather than a TOML value parser. It
-// finds raw table extents without mistaking quoted `]` characters or
-// table-looking lines inside multiline strings for structural delimiters. Any
-// unfinished lexical state makes provider replacement unsafe, so signed mode
-// refuses before either config or restore state is written.
-function tomlTableHeaders(contents) {
-  const lines = contents.split("\n");
-  const headers = [];
-  let multiline;
-  let arrayDepth = 0;
-
-  for (let lineIndex = 0; lineIndex < lines.length; lineIndex += 1) {
-    const line = lines[lineIndex];
-    const lineNumber = lineIndex + 1;
-    if (!multiline && arrayDepth === 0) {
-      const header = tableHeaderAtLine(line, lineNumber);
-      if (header) {
-        headers.push({ index: lineIndex, path: header });
-        continue;
-      }
-    }
-
-    let index = 0;
-    while (index < line.length) {
-      if (multiline) {
-        const quote = multiline === "basic" ? '"' : "'";
-        if (multiline === "basic" && line[index] === "\\") {
-          index += 2;
-          continue;
-        }
-        if (line[index] !== quote) {
-          index += 1;
-          continue;
-        }
-        let run = 1;
-        while (line[index + run] === quote) run += 1;
-        if (run >= 3) multiline = undefined;
-        index += run;
-        continue;
-      }
-
-      const character = line[index];
-      if (character === "#") break;
-      if (character === '"' || character === "'") {
-        const quote = character;
-        let run = 1;
-        while (line[index + run] === quote) run += 1;
-        if (run >= 3) {
-          multiline = quote === '"' ? "basic" : "literal";
-          index += 3;
-          continue;
-        }
-        index += 1;
-        let closed = false;
-        while (index < line.length) {
-          if (quote === '"' && line[index] === "\\") {
-            index += 2;
-            continue;
-          }
-          if (line[index] === quote) {
-            index += 1;
-            closed = true;
-            break;
-          }
-          index += 1;
-        }
-        if (!closed) {
-          ambiguousTomlBoundary(lineNumber, "a single-line string is unterminated");
-        }
-        continue;
-      }
-      if (character === "[") arrayDepth += 1;
-      else if (character === "]") {
-        if (arrayDepth === 0) {
-          ambiguousTomlBoundary(lineNumber, "an unmatched closing bracket was found");
-        }
-        arrayDepth -= 1;
-      }
-      index += 1;
-    }
-  }
-
-  if (multiline) {
-    ambiguousTomlBoundary(lines.length, `an unterminated multiline ${multiline} string was found`);
-  }
-  if (arrayDepth !== 0) {
-    ambiguousTomlBoundary(lines.length, "an unterminated array was found");
-  }
-  return { lines, headers };
-}
-
 function providerTableRanges(contents, providerId) {
-  const { lines, headers } = tomlTableHeaders(contents);
+  const { lines, headers } = scanTomlDocument(contents);
   const starts = headers.filter(({ path: header }) =>
     header[0] === "model_providers" && header[1] === providerId
   );
